requirements.txt
PySocks>=1.5.7
requests>=2.9.0